backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
google-genai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
requests
